requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
crispy-bootstrap5==0.7
Django==4.2
django-appconf==1.0.5
django-countries==7.5.1
django-crispy-forms==2.0
django-phonenumber-field==7.1.0
django-phonenumbers==1.0.1
django-user-accounts==3.2.0
idna==3.4
install==1.3.5
phonenumbers==8.13.11
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
